epocs log file is now named YYYYMMDD-HHMMSS-python.txt and it is accompanied by a matching YYYYMMDD-HHMMSS-operator.txt
ReflexConditioningSignalProcessing recompiled to be less verbose (sticking with
bci2000.org rHEAD sources for this module)

Operator.exe and OperatorLib.dll now recompiled from sources rolled back to
bci2000.org  r4528 (but with patches to recreate the new --SystemLogFile option
and to fix the $YYYYMMDD bug).  The rollback is in the hope that the sporadic
crashes will no longer happen.